creaet buttonsComponent in order to make code cleaner
