+++ src/components/commonComponents/ButtonsComponents/ButtonComponent.css
@@ -1,0 +1,14 @@
+.buttons {
+  display: flex;
+  justify-content: space-evenly;
+  margin-top: 8px;
+  margin-inline: 5px;
+}
+
+.orText {
+  display: flex;
+  font-size: medium;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+}
